Saprykin_Dmitry/requirements.txt
promptflow
promptflow-tools
azure-search-documents
pytest
langchain
langchain-community
unstructured
markdown
openai
